+++ src/Web/Controllers/OrderController.cs
@@ -29,7 +29,7 @@ public class OrderController : Controller
     }
 
     [HttpGet("{orderId}")]
-    public async Task<IActionResult> Detail(int orderId)
+    public async Task<IActionResult> Detail(long orderId)
     {
         Guard.Against.Null(User?.Identity?.Name, nameof(User.Identity.Name));
         var viewModel = await _mediator.Send(new GetOrderDetails(User.Identity.Name, orderId));
