+++ src/Web/Features/OrderDetails/GetOrderDetails.cs
@@ -6,9 +6,9 @@ namespace Microsoft.eShopWeb.Web.Features.OrderDetails;
 public class GetOrderDetails : IRequest<OrderDetailViewModel>
 {
     public string UserName { get; set; }
-    public int OrderId { get; set; }
+    public long OrderId { get; set; }
 
-    public GetOrderDetails(string userName, int orderId)
+    public GetOrderDetails(string userName, long orderId)
     {
         UserName = userName;
         OrderId = orderId;
